+++ trinity/common/models/__init__.py
@@ -43,24 +43,14 @@ def create_inference_models(
     from ray.util.placement_group import placement_group, placement_group_table
     from ray.util.scheduling_strategies import PlacementGroupSchedulingStrategy
 
-    from trinity.common.models.vllm_async_model import vLLMAysncRolloutModel
     from trinity.common.models.vllm_model import vLLMRolloutModel
 
     engine_num = config.explorer.rollout_model.engine_num
     tensor_parallel_size = config.explorer.rollout_model.tensor_parallel_size
 
-    if (
-        config.explorer.rollout_model.enable_openai_api
-        and config.explorer.rollout_model.engine_type != "vllm_async"
-    ):
-        raise ValueError("OpenAI API is only supported for vllm_async engine")
-
     rollout_engines = []
-
-    if config.explorer.rollout_model.engine_type == "vllm":
+    if config.explorer.rollout_model.engine_type.startswith("vllm"):
         engine_cls = vLLMRolloutModel
-    elif config.explorer.rollout_model.engine_type == "vllm_async":
-        engine_cls = vLLMAysncRolloutModel
     else:
         raise ValueError(f"Unknown engine type: {config.explorer.rollout_model.engine_type}")
 
@@ -122,10 +112,10 @@ def create_inference_models(
         for _ in range(model_config.engine_num):
             bundles_for_engine = allocator.allocate(model_config.tensor_parallel_size)
             model_config.enable_openai_api = True
-            model_config.engine_type = "vllm_async"
+            model_config.engine_type = "vllm"
             model_config.bundle_indices = ",".join([str(bid) for bid in bundles_for_engine])
             engines.append(
-                ray.remote(vLLMAysncRolloutModel)
+                ray.remote(vLLMRolloutModel)
                 .options(
                     num_cpus=0,
                     num_gpus=0 if model_config.tensor_parallel_size > 1 else 1,
